GPRS收到铱星设备的数据但是Web服务没有收到的原因是height字段的数据类型不一致，同时完善日志调试信息。

+++ src/main/java/com/zistone/gprs/repository/DeviceInfoRepository.java
@@ -24,7 +24,7 @@ public interface DeviceInfoRepository extends JpaRepository<DeviceInfo, Integer>
     @Query("update DeviceInfo device set device.m_state = 1,device.m_lat = :lat,device.m_lot = :lot,device.m_height = :height,device" +
             ".m_updateTime = CURRENT_TIMESTAMP" + " where " + "device.m_deviceId = " + ":deviceId")
     int UpdateLocationByDeviceId(
-            @Param("deviceId") String deviceId, @Param("lat") double lat, @Param("lot") double lot, @Param("height") double height);
+            @Param("deviceId") String deviceId, @Param("lat") double lat, @Param("lot") double lot, @Param("height") int height);
 
     @Transactional
     //清除底层持久化上下文
@@ -35,7 +35,7 @@ public interface DeviceInfoRepository extends JpaRepository<DeviceInfo, Integer>
             @Param("deviceId") String deviceId,
             @Param("lat") double lat,
             @Param("lot") double lot,
-            @Param("height") double height, @Param("temperature") int temperature, @Param("electricity") int electricity);
+            @Param("height") int height, @Param("temperature") int temperature, @Param("electricity") int electricity);
 
     @Transactional
     //清除底层持久化上下文
